Proof of concept for funnel api
